+++ HpcConnectionFramework/SchedulerAdapters/LinuxLocal/Generic/LinuxLocalSchedulerAdapter.cs
@@ -1,3 +1,12 @@
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Runtime;
+using System.Text;
+using System.Threading.Tasks;
+using log4net;
+using Renci.SshNet;
 using HEAppE.DomainObjects.ClusterInformation;
 using HEAppE.DomainObjects.JobManagement;
 using HEAppE.DomainObjects.JobManagement.JobInformation;
@@ -7,14 +16,6 @@ using HEAppE.HpcConnectionFramework.SchedulerAdapters.Interfaces;
 using HEAppE.HpcConnectionFramework.SystemCommands;
 using HEAppE.HpcConnectionFramework.SystemConnectors.SSH;
 using HEAppE.HpcConnectionFramework.SystemConnectors.SSH.DTO;
-using log4net;
-using Renci.SshNet;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Runtime;
-using System.Text;
 
 namespace HEAppE.HpcConnectionFramework.SchedulerAdapters.Generic.LinuxLocal;
 
@@ -380,6 +381,58 @@ public class LinuxLocalSchedulerAdapter : ISchedulerAdapter
     public bool MoveJobFiles(object schedulerConnectionConnection, SubmittedJobInfo jobInfo, IEnumerable<Tuple<string, string>> sourceDestinations)
     {
         return _commands.CopyJobFiles(schedulerConnectionConnection, jobInfo, sourceDestinations);
+    }
+
+    public async Task<dynamic> CheckClusterAuthenticationCredentialsStatus(object connectorClient, ClusterProjectCredential clusterProjectCredential, ClusterProjectCredentialCheckLog checkLog)
+    {
+        SshCommandWrapper command;
+        Cluster cluster = clusterProjectCredential.ClusterProject.Cluster;
+
+        int clusterConnectionFailedCount = 0;
+        int dryRunJobFailedCount = 0;
+
+        foreach (var nodeType in cluster.NodeTypes)
+        {
+            var partition = nodeType.Queue;
+            var script_name = $"{_scripts.LinuxLocalCommandScriptPathSettings.ScriptsBasePath}/{_linuxLocalCommandScripts.RunLocalCmdScriptName}";
+            var testCommand = $"[ -f {script_name} ]"; // just check that file to run scripts exists
+            var sshCommand = $"{_commands.InterpreterCommand} eval `(" + testCommand + ")`";
+            sshCommand = sshCommand.Replace("\r\n", "\n").Replace("\r", "\n");
+            try
+            {                
+                command = SshCommandUtils.RunSshCommand(new SshClientAdapter((SshClient)connectorClient), sshCommand);
+                checkLog.VaultCredentialOk = true;
+                checkLog.ClusterConnectionOk = true;
+                if (command.ExitStatus == 0)
+                {
+                    checkLog.DryRunJobOk = true;
+                }
+                else
+                {
+                    checkLog.DryRunJobOk = false;
+                    checkLog.ErrorMessage += command.Error + "\n";
+                        ++dryRunJobFailedCount;
+                }
+            }
+            catch (SshCommandException e)
+            {
+                ++clusterConnectionFailedCount;
+                checkLog.ErrorMessage += e.Message + "\n";
+            }
+            catch (Exception e)
+            {
+                checkLog.ErrorMessage += e.Message + "\n";
+            }
+        }
+
+        if (clusterConnectionFailedCount > 0)
+            checkLog.ClusterConnectionOk = false;
+
+        if (dryRunJobFailedCount > 0)
+            checkLog.DryRunJobOk = false;
+
+        await Task.Delay(1);
+        return null;
     }
 
     #endregion